User weight calibration
The user is now able to calibrate his weight. The yellow LED will blink after installing the system, which means the user should lay down on his bed. After the user layed down the yellow LED stops blinking and the user should now wait for five seconds. After five second the yellow LED will turn off, the blue LED will turn on (which means the systems may lock the door once again) and the system is calibrated to his user. To recalibrate the system, just hold the lock (the button in this prototype) for 1,5 seconds, whereafter the recalibration starts.